+++ Package.swift
@@ -5,6 +5,7 @@ import PackageDescription
 
 let package = Package(
     name: "CountdownLabel",
+    platforms: [.iOS(.v11)],
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
@@ -20,9 +21,7 @@ let package = Package(
         // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
             name: "CountdownLabel",
-            dependencies: []),
-        .testTarget(
-            name: "CountdownLabelTests",
-            dependencies: ["CountdownLabel"]),
+            dependencies: [])
+        
     ]
 )
