+++ Sources/CountdownLabel/Info.plist
