+++ Sources/CountdownLabel/LTMorphingLabel/Info.plist
