+++ Sources/CountdownLabel/LTMorphingLabel/LTCharacterDiffResult.swift
