+++ Sources/CountdownLabel/LTMorphingLabel/LTEasing.swift
