+++ Sources/CountdownLabel/LTMorphingLabel/LTEmitterView.swift
