+++ Sources/CountdownLabel/LTMorphingLabel/LTMorphingEffect.swift
