+++ Sources/CountdownLabel/LTMorphingLabel/LTMorphingLabel+Anvil.swift
