+++ Sources/CountdownLabel/LTMorphingLabel/LTMorphingLabel+Burn.swift
