+++ Sources/CountdownLabel/LTMorphingLabel/LTMorphingLabel+Evaporate.swift
