+++ Sources/CountdownLabel/LTMorphingLabel/LTMorphingLabel+Pixelate.swift
