+++ Sources/CountdownLabel/LTMorphingLabel/LTMorphingLabel+Sparkle.swift
