+++ Sources/CountdownLabel/LTMorphingLabel/LTStringDiffResult.swift
